Split out generic WalkDirectoryForFiles functionality
